requirements.txt
app-version==0.2.1
dj-database-url==0.4.2
Django==1.10.5
django-appconf==1.0.2
django-inspectional-registration==0.6.1
gunicorn==19.6.0
psycopg2==2.6.2
tolerance==0.1.2
